dnssec-validity-check/requirements.txt
dnspython
#pycryptodome
pycryptodomex
ecdsa

